CI: use flutter_tools for collecting coverage
Same change as canonical/dbus.dart#337 but because of different
`test_cov` breakage. It doesn't work with the latest stable Dart SDK
SDK version 2.17. It throws an exception in the CI, and hangs locally.

https://github.com/canonical/fwupd.dart/runs/6398669148

NOTE: this will only use Flutter tooling for collecting test coverage
in the CI without introducing an actual dependency to the Flutter SDK
on the package level.